nov-14

+++ admin/scripts/updateDBFeature.php
@@ -76,23 +76,25 @@ switch ($_POST[action]) {
 			`dateUpdate` = now(),
 			`userUpdate` = ''
 		WHERE `feature`.`id` = '".$_POST[idFeature]."' ;";
-		//mysql_query($sqlFeature) or die (mysql_error().$sqlFeature." error #80");
+		mysql_query($sqlFeature) or die (" error #80");
 		
 		//serie
 		$sqlDelSeries = "delete from `serieFeature` WHERE `idFeature` = '".$_POST[idFeature]."'";
-		//mysql_query($sqlDelSeries) or die (mysql_error()." error #85");
+		mysql_query($sqlDelSeries) or die (mysql_error()." error #85");
 		for ($i=0;$i<$_POST[lengthSerie];$i++){
 			$serieOpt = "rdSerie".$i;
 			$serieName = "txtSerie".$i;
 			if ($i > 0) { $valuesSerieInput .= ","; }
 			$valuesSerieInput .= "(NULL, '".$_POST[idFeature]."', '".$_POST[$serieName]."', '".$_POST[$serieOpt]."', now(), now(), NULL)";
 		}
-		$sqlAddSeries = "insert into `serieFeature` (`id`, `idFeature`, `description`, `option`, `dateCreate`, `dateUpdate`, `userUpdate`) VALUES ".$valuesSerieInput;
-		//mysql_query($sqlAddSeries) or die (" error #93");
+		if ($valuesSerieInput != ""){
+			$sqlAddSeries = "insert into `serieFeature` (`id`, `idFeature`, `description`, `option`, `dateCreate`, `dateUpdate`, `userUpdate`) VALUES ".$valuesSerieInput;
+			mysql_query($sqlAddSeries) or die (" error #93");
+		}
 
 		//options
 		$sqlDelOpts = "delete from `optionsFeature` WHERE `idFeature` = '".$_POST[idFeature]."'";
-		//mysql_query($sqlDelOpts) or die (mysql_error()." error #85");
+		mysql_query($sqlDelOpts) or die (" error #100");
 		$o=0;
 		for ($i=0;$i<$_POST[lengthOptions];$i++){
 			$optIdOption = "txtOpt".$i;
@@ -103,14 +105,14 @@ switch ($_POST[action]) {
 				$o++;
 			}
 		}
-		$sqlAddOpts = "insert into `optionsFeature` (`id`, `idFeature`, `idOption`, `option`, `dateCreate`, `dateUpdate`, `userUpdate`) VALUES ".$valuesOptInput;
-		//mysql_query($sqlAddOpts) or die (mysql_error()." error #93");
-		echo $sqlAddOpts."<br>";
+		if ($valuesOptInput != ""){
+			$sqlAddOpts = "insert into `optionsFeature` (`id`, `idFeature`, `idOption`, `option`, `dateCreate`, `dateUpdate`, `userUpdate`) VALUES ".$valuesOptInput;
+			mysql_query($sqlAddOpts) or die (" error #110");
+		}
 
 		//color
 		$sqlDelColor = "delete from `colorModel` where `idModel` = '".$_POST[idModel]."'";
-		//mysql_query($sqlDelColor) or die (mysql_error()." error #96");
-		//echo $sqlDelColor."<br>";
+		mysql_query($sqlDelColor) or die (mysql_error()." error #115");
 		for ($i=0;$i<$_POST[colorLength];$i++){
 			$colorName = $_POST["colorInputName".$i];
 			$colorApp = $_POST["colorInputApp".$i];
@@ -119,9 +121,11 @@ switch ($_POST[action]) {
 			if ($i > 0) { $valuesColorInput .= ","; }
 			$valuesColorInput .= "(NULL, '".$_POST[idModel]."', '".$_POST["colorInputName".$i]."', '".$_POST["colorInputColor".$i]."', '".$_POST["colorInputApp".$i]."', '".$_POST["colorInputType".$i]."', now(), now(), NULL)";
 		}
-		$sqlAddColor = "insert into `colorModel` (`id`, `idModel`, `name`, `hexa`, `application`, `type`, `dateCreate`, `dateUpdate`, `userUpdate`) VALUES ".$valuesColorInput;
-		//mysql_query($sqlAddColor) or die (mysql_error().$sqlAddColor." error #106");
-		//echo $sqlAddColor."<br>";
+		if ($valuesColorInput != ""){
+			$sqlAddColor = "insert into `colorModel` (`id`, `idModel`, `name`, `hexa`, `application`, `type`, `dateCreate`, `dateUpdate`, `userUpdate`) VALUES ".$valuesColorInput;
+			mysql_query($sqlAddColor) or die (" error #126");
+		}
+
 
 		//pictures
